Python script to determine the sentiment of twitter mentions
This python script is the main submission for EC601 Project 2. This python script calls the twitter API to find the mentions of any given twitter handle across any given range of dates. The google NLP API is then called to determine the sentiment score of the found tweets. The script then translates the found sentiment score to a meaningful sentiment analysis output to the user.