EC601 Project 2 Product Submission
This python program searches twitter for mentions of a specific user and determines the sentiment of the found tweets using the google NLP API. The program prompts users for a twitter handle, as well as a start time and an end time to determine the allowable range of timestamps the search returns. This program can find up to ~800 tweets at once and determines the sentiment for all the tweets at once and translates the google sentiment score to a relevant statement regarding sentiment.